FDI-case-study-using-Tableau-Public
In this Case Study, we will explore the FDI dataset and analyse it from the perspective of the business. We shall look at various sectors to determine which sector will prove most beneficial to collaborate with, for the company.

FDI : A foreign direct investment (FDI) is an investment made by a corporation of a certain country into a corporation of another country. However, FDI is different form indirect investments. If a foreign investor passively invests in an Indian company by owning stocks or bonds, this will not be considered as FDI. But, if a foreign investor invests in an Indian company, and controls technology, management and other aspects, this will be considered as foreign direct investment. For example, a foreign institution invests in a company that is listed on a major stock exchange. However, such institutions do not control various aspects of management in the firms in which they invest in. For an investment to be considered FDI, the firm you invest in should be located in a foreign jurisdiction and control ownership to a degree. Often, it is the large firms that engage in FDI.

Which sector receives the maximum direct funding for a given year?

Which sector receives the minimum direct funding for a given year?

Is there any trend of overall foreign direct investment(filter to select any indivual sector)

Which sectors reported high growth in FDI in last 5 years?

Which sectors reported lowest growth in FDI in last 5 years?

Which sectors reported the most variation overall?

What is the proportion of each sector in FDI?

Are there some specific clusters that are present here?

What is the forecast for next year?